requirements.txt
python-dotenv==0.15.0
python-telegram-bot==13.0
pytest==6.1.2
requests==2.24.0